dockerFlaskRun/requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorlog==6.7.0
cryptography==39.0.2
DBUtils==3.0.3
ddt==1.2.1
dnspython==2.3.0
exceptiongroup==1.1.1
Flask==2.2.3
Flask-Cors==3.0.10
gevent==22.10.2
greenlet==2.0.2
HTMLReport==2.1.5
idna==3.4
importlib-metadata==6.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpath==0.82
lxml==4.8.0
MarkupSafe==2.1.2
mysql-connector==2.2.9
mysql-connector-python-rf==2.2.2
mysqlclient==2.1.1
numpy==1.21.4
packaging==23.1
pluggy==1.0.0
pycparser==2.21
pycryptodome==3.11.0
pymongo==4.3.3
PyMySQL==0.9.3
pymysql-pool==0.4.3
PyMysqlPool==1.0.9
pyOpenSSL==23.0.0
pytest==7.3.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
requests-mock==1.10.0
requests-toolbelt==0.9.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
selenium==3.141.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.25.3
Werkzeug==2.2.3
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
